+++ CHANGELOG.markdown
@@ -12,7 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
   * The optimization now uses bisection search to locate the proper segment when performing Jump Table compression, providing up to 15% speedup in some scenarios.
   * The jump table can be padded not just towards zero, but also towards the maximum value, to remove the need for the range check. Requires `target-optimization` to be set to `specific`. 
   * The `case-optimization-strength` option now has a range from `0` to `6`. Each additional value significantly increases the number of segment arrangements considered, as well as optimization time.
-  * When compressing jump tables, the optimizer also generates segment configuration for a full bisectional search. This may improve case expressions too small for a full jump table optimization.
+  * When compressing jump tables, the optimizer also generates segment configuration for a full bisection search. This may improve case expressions too small for a full jump table optimization.
+  * The bodies of `when` branches are moved into correct places inside the case expression when possible, to avoid unnecessary jumps.
   * The performance of the Case Switching optimization (in terms of compilation time) has been significantly improved.  
 * It is now possible to turn off [Jump Table Compression](/doc/syntax/SYNTAX-6-OPTIMIZATIONS.markdown#jump-table-compression) by setting the `case-optimization-strength` compiler option to `0`.